requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.10
requests==2.32.3
Jinja2==3.1.4
python-dotenv==1.0.1
